Events reorganization
