+++ src/events/mod.rs
@@ -1,0 +1,4 @@
+pub mod auth;
+pub mod group;
+pub mod message;
+pub mod user;
